refactor functions into separate files
