+++ docs/generator/src/find.js
@@ -1,0 +1,23 @@
+const glob = require('glob').sync;
+const readFile = require('fs').readFileSync;
+const { parse } = require('@babel/parser');
+
+const DOCUMENTATION_IDENTIFIER = '@Documented';
+const blockComments = (ast) => ast.comments.filter((node) => node.type === 'CommentBlock');
+
+function findDocumentationComment(fileContent) {
+  return blockComments(parse(fileContent))
+    .map((comment) => comment.value)
+    .find((content) => content.includes(DOCUMENTATION_IDENTIFIER));
+}
+
+// Returns [path, content]
+function findDocumentedFiles(basePath, globPattern) {
+  const filepaths = glob(globPattern, {
+    cwd: basePath
+  });
+
+  return filepaths.map((filePath) => [filePath, readFile(filePath, 'utf-8')]).filter(([_, content]) => findDocumentationComment(content));
+}
+
+module.exports = { findDocumentedFiles, findDocumentationComment };
